Refactor triangle-mode
Minor

Change `assert_supported` exception type from `ValueError` to `AssertionError`

Remove unused return-value

Remove unused code

Refactor single-trianglre mode as well

Remove unneeded assertion

Cosmetic

Fix single-triangle mode

Fix single-triangle mode

Refactor the b3-two-hop-triangle mode as well

Remove the importing of `T`

Remove unused code

Remove single-triangle mode, and fix a bunch of problems

Cleanup `calculate_profit` related code

Fix test 042

Fix  test 042

Remove unused code

Some more general cleanup

Cleanup

Remove unused code

Fix test 063

Cleanup

Cleanup

Fix tests

Cleanup

Cleanup

Fix test 042

Remove single_paiwise and multi_paiwise arb modes (major change)

Fix test 045

Fix test 045

Refactor the pairwise arb-mode classes

Ongoing refactor

Refactor triangle arb-mode classes

Cleanup

Remove file `scan_log_errors.py` from the repository

Cleanup

Remove unused code

Code reuse

Fix functions `create_sort_order` and `sort_key`
