+++ fastlane_bot/modes/base_pairwise.py
@@ -8,53 +8,55 @@ Defines the base class for pairwise arbitrage finder modes
 All rights reserved.
 Licensed under MIT.
 """
-import abc
-import itertools
-from typing import List, Tuple, Any, Union
+from typing import List, Tuple, Union
 
-from fastlane_bot.modes.base import ArbitrageFinderBase
 from fastlane_bot.tools.cpc import CPCContainer
-
+from fastlane_bot.tools.optimizer import PairOptimizer
+from fastlane_bot.modes.base import ArbitrageFinderBase
 
 class ArbitrageFinderPairwiseBase(ArbitrageFinderBase):
-    """
-    Base class for pairwise arbitrage finder modes
-    """
+    def find_arbitrage(self) -> Union[List, Tuple]:
+        all_tokens, combos = self.get_combos(self.flashloan_tokens, self.CCm)
+        if self.result == self.AO_TOKENS:
+            return all_tokens, combos
 
-    @abc.abstractmethod
-    def find_arbitrage(self, candidates: List[Any] = None, ops: Tuple = None, best_profit: float = 0, profit_src: float = 0) -> Union[List, Tuple]:
-        """
-        see base.py
-        """
-        pass
+        candidates = []
+        best_profit = 0
+        ops = None
 
-    @staticmethod
-    def get_combos(
-        CCm: CPCContainer, flashloan_tokens: List[str]
-    ) -> Tuple[List[Any], List[Any]]:
-        """
-        Get combos for pairwise arbitrage
+        for tkn0, tkn1 in combos:
+            CC = self.CCm.bypairs(f"{tkn0}/{tkn1}")
+            if len(CC) < 2:
+                continue
 
-        Parameters
-        ----------
-        CCm : CPCContainer
-            Container for all the curves
-        flashloan_tokens : list
-            List of flashloan tokens
+            for curve_combo in self.get_curve_combos(CC):
+                src_token = tkn1
+                if len(curve_combo) < 2:
+                    continue
+                try:
+                    CC_cc = CPCContainer(curve_combo)
+                    O = PairOptimizer(CC_cc)
+                    pstart = {tkn0: CC_cc.bypairs(f"{tkn0}/{tkn1}")[0].p}
+                    r = O.optimize(src_token, params=dict(pstart=pstart))
+                    trade_instructions_dic = r.trade_instructions(O.TIF_DICTS)
+                    trade_instructions_df = r.trade_instructions(O.TIF_DFAGGR)
+                    trade_instructions = r.trade_instructions()
+                except Exception as e:
+                    self.ConfigObj.logger.debug(f"[base_pairwise] {e}")
+                    continue
+                if trade_instructions_dic is None or len(trade_instructions_dic) < 2:
+                    # Failed to converge
+                    continue
+                # Update the results
+                best_profit, ops = self.update_results(
+                    src_token,
+                    r,
+                    trade_instructions_dic,
+                    trade_instructions_df,
+                    trade_instructions,
+                    candidates,
+                    best_profit,
+                    ops,
+                )
 
-        Returns
-        -------
-        all_tokens : list
-            List of all tokens
-
-        """
-        all_tokens = CCm.tokens()
-        flashloan_tokens_intersect = all_tokens.intersection(set(flashloan_tokens))
-        combos = [
-            (tkn0, tkn1)
-            for tkn0, tkn1 in itertools.product(all_tokens, flashloan_tokens_intersect)
-            # tkn1 is always the token being flash loaned
-            # note that the pair is tkn0/tkn1, ie tkn1 is the quote token
-            if tkn0 != tkn1
-        ]
-        return all_tokens, combos
+        return candidates if self.result == self.AO_CANDIDATES else ops
