+++ fastlane_bot/modes/triangle_multi.py
@@ -1,5 +1,5 @@
 """
-Defines the Triangular arbitrage finder class
+Defines the multi-triangle arbitrage finder class
 
 [DOC-TODO-OPTIONAL-longer description in rst format]
 
@@ -8,79 +8,124 @@ Defines the Triangular arbitrage finder class
 All rights reserved.
 Licensed under MIT.
 """
-from typing import List, Any, Tuple, Union
+from typing import List, Any, Tuple
+from itertools import product
 
 from fastlane_bot.modes.base_triangle import ArbitrageFinderTriangleBase
-from fastlane_bot.tools.cpc import CPCContainer
-from fastlane_bot.tools.optimizer import MargPOptimizer
-
 
 class ArbitrageFinderTriangleMulti(ArbitrageFinderTriangleBase):
-    """
-    Triangular arbitrage finder mode
-    """
-
     arb_mode = "multi_triangle"
 
-    def find_arbitrage(self, candidates: List[Any] = None, ops: Tuple = None, best_profit: float = 0, profit_src: float = 0) -> Union[List, Tuple]:
+    def handle_exchange(self):
+        assert self.base_exchange in self.ConfigObj.CARBON_V1_FORKS, "base_exchange for `multi_triangle` mode must be a carbon_v1 fork"
+
+    def get_combos(self, flashloan_tokens: List[str], CCm: Any) -> Tuple[List[str], List[Any]]:
         """
-        see base.py
+        Get combos for triangular arbitrage
+
+        Parameters
+        ----------
+        flashloan_tokens : list
+            List of flashloan tokens
+        CCm : object
+            CCm object
+
+        Returns
+        -------
+        combos : list
+            List of combos
+
         """
-
-        if self.base_exchange != "carbon_v1":
-            raise ValueError("base_exchange must be carbon_v1 for `multi` mode")
-
-        if candidates is None:
-            candidates = []
-
-        combos = self.get_combos(self.flashloan_tokens, self.CCm, arb_mode=self.arb_mode)
-
-        for src_token, miniverse in combos:
-            try:
-                CC_cc = CPCContainer(miniverse)
-                O = MargPOptimizer(CC_cc)
-                pstart = self.build_pstart(CC_cc, CC_cc.tokens(), src_token)
-                r = O.optimize(src_token, params=dict(pstart=pstart))
-                trade_instructions_dic = r.trade_instructions(O.TIF_DICTS)
-                if trade_instructions_dic is None or len(trade_instructions_dic) < 3:
-                    # Failed to converge
+        combos = []
+        all_base_exchange_curves = CCm.byparams(exchange=self.base_exchange).curves
+        for flt in flashloan_tokens:  # may wish to run this for one flt at a time
+            non_flt_base_exchange_curves = [
+                x for x in all_base_exchange_curves if flt not in x.pair
+            ]
+            for non_flt_base_exchange_curve in non_flt_base_exchange_curves:
+                target_tkny = non_flt_base_exchange_curve.tkny
+                target_tknx = non_flt_base_exchange_curve.tknx
+                base_exchange_curves = (
+                    CCm.bypairs(f"{target_tknx}/{target_tkny}")
+                    .byparams(exchange=self.base_exchange)
+                    .curves
+                )
+                if len(base_exchange_curves) == 0:
                     continue
-                trade_instructions_df = r.trade_instructions(O.TIF_DFAGGR)
-                trade_instructions = r.trade_instructions()
 
-            except Exception as e:
-                self.ConfigObj.logger.info(f"[triangle multi] {e}")
-                continue
-            profit_src = -r.result
+                base_direction_pair = base_exchange_curves[0].pair
+                base_direction_one = [curve for curve in base_exchange_curves if curve.pair == base_direction_pair]
+                base_direction_two = [curve for curve in base_exchange_curves if curve.pair != base_direction_pair]
 
-            # Get the cids
-            cids = [ti["cid"] for ti in trade_instructions_dic]
+                y_match_curves = CCm.bypairs(
+                    set(CCm.filter_pairs(onein=target_tknx))
+                    & set(CCm.filter_pairs(onein=flt))
+                )
+                x_match_curves = CCm.bypairs(
+                    set(CCm.filter_pairs(onein=target_tkny))
+                    & set(CCm.filter_pairs(onein=flt))
+                )
 
-            # Calculate the profit
-            profit = self.calculate_profit(src_token, profit_src, self.CCm, cids)
-            if str(profit) == "nan":
-                self.ConfigObj.logger.debug("profit is nan, skipping")
-                continue
+                y_match_curves_not_carbon = [
+                    x
+                    for x in y_match_curves
+                    if x.params.exchange != self.base_exchange
+                ]
+                if len(y_match_curves_not_carbon) == 0:
+                    continue
+                x_match_curves_not_carbon = [
+                    x
+                    for x in x_match_curves
+                    if x.params.exchange != self.base_exchange
+                ]
+                if len(x_match_curves_not_carbon) == 0:
+                    continue
+                if len(base_direction_one) > 0:
+                    get_miniverse(
+                        y_match_curves_not_carbon,
+                        base_direction_one,
+                        x_match_curves_not_carbon,
+                        flt,
+                        combos,
+                    )
+                if len(base_direction_two) > 0:
+                    get_miniverse(
+                        y_match_curves_not_carbon,
+                        base_direction_two,
+                        x_match_curves_not_carbon,
+                        flt,
+                        combos,
+                    )
+        return combos
 
-            # Handle candidates based on conditions
-            candidates += self.handle_candidates(
-                best_profit,
-                profit,
-                trade_instructions_df,
-                trade_instructions_dic,
-                src_token,
-                trade_instructions,
-            )
+def get_miniverse(
+    y_match_curves_not_carbon: List[Any],
+    base_exchange_curves: List[Any],
+    x_match_curves_not_carbon: List[Any],
+    flt: str,
+    combos: List[Any],
+):
+    """
+    Get miniverse for triangular arbitrage
 
-            # Find the best operations
-            best_profit, ops = self.find_best_operations(
-                best_profit,
-                ops,
-                profit,
-                trade_instructions_df,
-                trade_instructions_dic,
-                src_token,
-                trade_instructions,
-            )
+    Parameters
+    ----------
+    y_match_curves_not_carbon : list
+        List of curves that match the y token and are not on carbon
+    base_exchange_curves : list
+        List of curves on the base exchange
+    x_match_curves_not_carbon : list
+        List of curves that match the x token and are not on carbon
+    flt : str
+        Flashloan token
+    combos : list
+        List of combos
 
-        return candidates if self.result == self.AO_CANDIDATES else ops
+    """
+    external_curve_combos = list(
+        product(y_match_curves_not_carbon, x_match_curves_not_carbon)
+    )
+    miniverses = [
+        base_exchange_curves + list(combo) for combo in external_curve_combos
+    ]
+    combos += list(zip([flt] * len(miniverses), miniverses))
