+++ fastlane_bot/tests/test_047_Randomizer.py
@@ -128,8 +128,6 @@ flashloan_tokens = bot.RUN_FLASHLOAN_TOKENS
 CCm = bot.get_curves()
 pools = db.get_pool_data_with_tokens()
 
-arb_mode = "multi"
-
 
 # ------------------------------------------------------------
 # Test      047
@@ -140,7 +138,7 @@ def test_test_randomizer():
 # ------------------------------------------------------------
     
     # +
-    arb_finder = bot._get_arb_finder("multi")
+    arb_finder = bot._get_arb_finder("multi_pairwise_all")
     finder = arb_finder(
                 flashloan_tokens=flashloan_tokens,
                 CCm=CCm,
