+++ fastlane_bot/tests/test_058_BalancerIntegration.py
@@ -136,8 +136,6 @@ flashloan_tokens = bot.RUN_FLASHLOAN_TOKENS
 CCm = bot.get_curves()
 pools = db.get_pool_data_with_tokens()
 
-arb_mode = "multi_pairwise"
-
 
 # ------------------------------------------------------------
 # Test      058
